Iteration 1 for CS449 Project
For this iteration I made the framework to build off.

- added links to buttons to switch between activities

- added a number pad to select the item to get history

-added information about business owner that is scrollable

- made a template for how the information will be pulled up from a
database once it is built.

- spent a week researching databases and talked to Prof Burris who told
me to hold off on it until we go over it in class

Prof Burris made a comment about using QR codes which I thought was good
at the time but then realized it will not work for what I would like it
to do in the end. I would like the app to be utilized by people who can
not visit the museum which would not be possible with only doing QR
codes.